hyperliquid-twap/requirements.txt
# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# AWS & Data Processing
boto3==1.34.34
pyarrow==15.0.0
pandas==2.2.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
black==24.1.1
ruff==0.1.14
